Install and setup certbot for SSL certificates
During the packer build, we need to modify the
DNS records to verify we own the domain we’re
registering certificates for.

There is a HTTP auth method that only looks at
the IP address (A record) for the subdomain, but
this won’t match until we allocate the Elastic IP
to the box which we shouldn’t do while building
the AMI.

If we did this, the Packer SSH session would fail
anyway.